WIP: Bash command line parsing, part 3: some renaming.
